requirements.txt
flask==3.0.2
scikit-learn==1.4.0
scikit-image==0.22.0
unzip==1.0.0
joblib==1.3.2
pandas==2.2.0
natsort==8.4.0
numpy==1.26.3
pillow==10.2.0